new update

+++ resources/views/admin/minum.blade.php
@@ -7,16 +7,20 @@
     <meta name="csrf-token" content="{{ csrf_token() }}">
 	<!-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" />  -->
 	<!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.min.css" /> -->
-    <link href="https://cdn.datatables.net/1.10.16/css/jquery.dataTables.min.css" rel="stylesheet">
+    <link href="https://cdn.datatables.net/1.10.16/css/jquery.dataTables.min.css" rel="stylesheet"> 
     <link href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css" rel="stylesheet">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>  
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.0/jquery.validate.js"></script>
     <script src="https://cdn.datatables.net/1.10.16/js/jquery.dataTables.min.js"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
     <script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap4.min.js"></script>
     <script src="https://cdn.datatables.net/buttons/1.0.3/js/dataTables.buttons.min.js"></script>
+    <script src="https://cdn.datatables.net/1.11.0/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/datetime/1.1.1/js/dataTables.dateTime.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.0/jquery.validate.js"></script>
     <link rel="stylesheet" href="https://cdn.datatables.net/buttons/1.0.3/css/buttons.dataTables.min.css">
-    <script src="{{ asset('vendor/datatables/buttons.server-side.js') }}"></script>
+    <!-- <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>   --> 
+    <!-- <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script> -->
+    <script src="https://code.jquery.com/jquery-3.5.1.js"></script>
+    <script src="{{ asset('vendor/datatables/buttons.server-side.js') }}"></script> 
+    <script src="../../extensions/Editor/js/dataTables.editor.min.js"></script>
 @stop
 
 @section('content_header')
@@ -50,7 +54,7 @@
                 <th>Jumlah Minum (l)</th>
                 <th>Nama Farm</th>
                 <th>Mitra</th>
-                <th>Aksi</th>
+                <th style="width:45px">Aksi</th>
             </tr>
         </thead>
         <tbody>
@@ -74,7 +78,7 @@
                     {data: 'nama_farm', name: 'nama_farm'},
                     {data: 'nama', name: 'nama'},
                     { data: 'minum_id', name: 'minum_id', render: function (data, type, row, meta) { 
-                        return '<a class="btn btn-info btn-sm mr-2" href="/admin/minum/'+row.minum_id+'/edit"><i class="fas fa-edit"></i></a><a class="btn btn-danger btn-sm" href="/admin/minum/'+row.minum_id+'/delete" ><i class="fas fa-trash"></i></a>';
+                        return '<a class="btn btn-info btn-sm mr-2" href="/admin/minum/'+row.minum_id+'/edit"><i class="fas fa-edit"></i></a><a class="btn btn-danger btn-sm" href="/admin/minum/'+row.minum_id+'/delete" onclick="return confirm(`Apakah data ini mau dihapus?`)"><i class="fas fa-trash"></i></a>';
                     }},
                 ]
             });
