+++ resources/views/admin/user.blade.php
@@ -6,18 +6,22 @@
     <link rel="stylesheet" href="/css/app.css"><link rel="shortcut icon" href="{{ asset('img/mardawavector.png') }}"/>
     <meta name="csrf-token" content="{{ csrf_token() }}">
 	<!-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" />  -->
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.min.css" />
+	<!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.min.css" /> -->
     <link href="https://cdn.datatables.net/1.10.16/css/jquery.dataTables.min.css" rel="stylesheet">
     <link href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css" rel="stylesheet">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>  
+    <!-- <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>   --> 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.0/jquery.validate.js"></script>
     <script src="https://cdn.datatables.net/1.10.16/js/jquery.dataTables.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
     <script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap4.min.js"></script>
     <script src="https://cdn.datatables.net/buttons/1.0.3/js/dataTables.buttons.min.js"></script>
     <link rel="stylesheet" href="https://cdn.datatables.net/buttons/1.0.3/css/buttons.dataTables.min.css">
-    <script src="{{ asset('vendor/datatables/buttons.server-side.js') }}"></script>
+    <script src="https://code.jquery.com/jquery-3.5.1.js"></script>  
+    <script src="{{ asset('vendor/datatables/buttons.server-side.js') }}"></script> 
+    <script src="../../extensions/Editor/js/dataTables.editor.min.js"></script>
 @stop
+
+@if(auth()->user()->isAdmin())
 
 @section('content_header')
     <h1 class="m-0 text-dark">Manajemen User</h1>
@@ -41,7 +45,7 @@
 
     <div class="row justify-content-between">
         <div class="col-md">
-            <a class="btn btn-success mr-2 mb-2" style="font-family: Source Sans Pro"; href="/admin/user/tambah"><i class="fas fa-plus"></i> Tambah Data</a>
+            <a class="btn btn-success ml-2 mr-2 mb-2" style="font-family: Source Sans Pro"; href="/admin/user/tambah"><i class="fas fa-plus"></i> Tambah Data</a>
             <a href="/admin/user/export_excel" class="btn btn-secondary mb-2" target="_blank"><i class="fas fa-print"></i> Export Excel</a>
         </div>
     </div>
@@ -55,7 +59,7 @@
                 <th>Nama</th>
                 <th>Email</th>
                 <th>Hak Akses</th>
-                <th>Aksi</th>
+                <th style="width:45px">Aksi</th>
             </tr>
         </thead>
         <tbody>
@@ -78,7 +82,7 @@
             {data: 'email', name: 'email'},
             {data: 'hak_akses', name: 'hak_akses'},
             { data: 'id', name: 'id', render: function (data, type, row, meta) { 
-                return '<a class="btn btn-info btn-sm mr-2" href="/admin/user/'+row.id+'/edit"><i class="fas fa-edit"></i></a><a class="btn btn-danger btn-sm" href="/admin/user/'+row.id+'/delete" ><i class="fas fa-trash"></i></a>';
+                return '<a class="btn btn-info btn-sm mr-2" href="/admin/user/'+row.id+'/edit"><i class="fas fa-edit"></i></a><a class="btn btn-danger btn-sm" href="/admin/user/'+row.id+'/delete" onclick="return confirm(`Apakah data ini mau dihapus?`)"><i class="fas fa-trash"></i></a>';
             }},
             
         ]
@@ -88,8 +92,12 @@
 </script>
 </div>
 @stop
+@endif
 
 @section('scripts')
 <script src="js/jquery.min.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/js/bootstrap.bundle.min.js"></script>
 
 @endsection
+
+
